feat: wire up stats push code
